feat: update react demo
