+++ examples/demo/package.json
@@ -17,11 +17,13 @@
     "@algolia/ui-components-horizontal-slider-react": "1.0.0",
     "@algolia/ui-components-horizontal-slider-theme": "1.0.0",
     "algoliasearch": "4.10.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "search-insights": "2.0.0"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "^6.10.0",
+    "search-insights": "2.4.0"
   },
   "devDependencies": {
+    "@types/react-dom": "^18.0.11",
     "parcel": "2.8.3"
   }
 }
