+++ examples/demo/src/components/BundleView/index.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useMemo, useState } from 'react';
 
-import { BaseObject, BundleViewProps } from './types';
+import { BaseObject, BundleViewTranslations } from '../../types';
 
-import './BundleView.css';
+import './style.css';
 
 function getAmountDefault<TObject extends BaseObject>(items: TObject[]) {
   return items.reduce((sum, current) => sum + current.price.value, 0);
@@ -12,10 +12,19 @@ function formatPriceDefault(price: number) {
   return `$${price}`;
 }
 
+type BundleViewProps<TObject> = {
+  currentItem: TObject;
+  formatPrice?: (price: number) => string;
+  getAmount?(items: TObject[]): number;
+  itemComponent({ item }): JSX.Element;
+  items: TObject[];
+  translations?: BundleViewTranslations;
+};
+
 export function BundleView<TObject extends BaseObject>(
   props: BundleViewProps<TObject>
 ): JSX.Element {
-  const items = [props.currentItem, ...props.items];
+  const items = Array.from(new Set([props.currentItem, ...props.items]));
   const formatPrice = props.formatPrice ?? formatPriceDefault;
   const getAmount = props.getAmount ?? getAmountDefault;
