+++ examples/demo/src/components/index.ts
@@ -1,0 +1,5 @@
+export * from './Autocomplete';
+export * from './BundleItem';
+export * from './BundleView';
+export * from './Facet';
+export * from './Hit';
