+++ examples/demo/src/config.ts
@@ -1,0 +1,3 @@
+export const appId = 'XX85YRZZMV';
+export const apiKey = '098f71f9e2267178bdfc08cc986d2999';
+export const indexName = 'test_FLAGSHIP_ECOM_recommend';
