+++ examples/demo/src/index.tsx
@@ -1,6 +1,35 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
-import App from './App';
+import { HomePage } from './routes/HomePage';
+import { ProductPage } from './routes/ProductPage';
+import { Root } from './routes/Root';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import '@algolia/autocomplete-theme-classic';
+import '@algolia/ui-components-horizontal-slider-theme';
+import './App.css';
+import './Recommend.css';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Root />,
+    children: [
+      {
+        path: '/',
+        element: <HomePage />,
+      },
+      {
+        path: '/product/:id',
+        element: <ProductPage />,
+      },
+    ],
+  },
+]);
+
+createRoot(document.getElementById('root')!).render(
+  <React.StrictMode>
+    <RouterProvider router={router} />
+  </React.StrictMode>
+);
