--- examples/demo/src/types/BundleItem.ts
@@ -1,4 +1,0 @@
-export type BundleItemProps<TObject> = {
-  item: TObject;
-  onSelect(item: TObject): void;
-};
