+++ examples/demo/src/types/BundleView.ts
@@ -9,12 +9,3 @@ export type BundleViewTranslations = Partial<{
   thisArticle: string;
   addToCart(count: number): string;
 }>;
-
-export type BundleViewProps<TObject> = {
-  currentItem: TObject;
-  formatPrice?: (price: number) => string;
-  getAmount?(items: TObject[]): number;
-  itemComponent({ item: TObject }): JSX.Element;
-  items: TObject[];
-  translations?: BundleViewTranslations;
-};
