+++ examples/demo/src/types/index.ts
@@ -1,5 +1,3 @@
-export * from './BundleItem';
 export * from './BundleView';
 export * from './FacetHit';
-export * from './Hit';
 export * from './ProductHit';
